backend/pyproject.toml
[project]
    dependencies = [
        "dj-database-url>=3.0.1",
        "django-cors-headers>=4.9.0",
        "django>=5.2.5",
        "djangorestframework-simplejwt>=5.5.1",
        "djangorestframework>=3.16.1",
        "google-auth>=2.40.3",
        "google-genai>=1.38.0",
        "gunicorn>=23.0.0",
        "psycopg[binary]>=3.2.10",
        "python-dotenv>=1.1.1",
        "whitenoise>=6.11.0",
    ]
    name = "nutriplan-backend"
    requires-python = ">=3.13"
    version = "0.5"

[dependency-groups]
    dev = [
        "factory-boy>=3.3.3",
        "faker>=37.11.0",
        "freezegun>=1.5.5",
        "pytest-cov>=7.0.0",
        "pytest-django>=4.11.1",
        "pytest>=8.4.2",
        "ruff>=0.12.12",
    ]

[tool.coverage.run]
    branch = true
    omit = [
        "**/__init__.py",
        "**/admin/*",
        "**/apps.py",
        "**/management/*",
        "**/migrations/*",
        "**/services/llm/*",
        "**/tests/*",
    ]
    source = ["nutriplan"]

[tool.coverage.report]
    exclude_lines = [
        "if TYPE_CHECKING:",
        "pass",
        "raise NotImplementedError",
    ]
    omit = [
        "**/__init__.py",
        "**/admin/*",
        "**/apps.py",
        "**/management/*",
        "**/migrations/*",
        "**/services/llm/*",
        "**/tests/*",
    ]
    show_missing = true
    skip_covered = true

[tool.pytest.ini_options]
    DJANGO_SETTINGS_MODULE = "backend.settings"
    addopts                = "-q --reuse-db --cov=nutriplan --cov-config=pyproject.toml --cov-report=term-missing"
    console_output_style   = "count"
    pythonpath             = "src"

[tool.ruff.lint]
    ignore = [
        "COM812",
        "D200",
        "D202",
        "D203",
        "D212",
        "D401",
    ]
    select = ["ALL"]

    [tool.ruff.lint.isort]
        case-sensitive             = true
        combine-as-imports         = true
        force-sort-within-sections = true
        order-by-type              = true
        relative-imports-order     = "closest-to-furthest"
        split-on-trailing-comma    = false

    [tool.ruff.lint.per-file-ignores]
        "**/admin/__init__.py" = ["F401"]
        "**/prompt*" = ["E501"]
        "**/migrations/*" = [
            "ANN001",
            "ARG001",
            "D100",
            "D101",
            "D103",
            "D104",
            "E501",
            "RUF012",
        ]
        "**/tests/*" = ["D100", "D101", "D103", "D104", "D106", "PT013", "S101"]
        "src/backend/settings.py" = ["E501"]
